* 5.4:
  [Process] intersect with getenv() in case-insensitive manner to get default envs
  [Serializer] fix support for lazy/unset properties
  Fix redundant type casts
  [Notifier] Fix AllMySms bridge body content
  Revert "[DoctrineBridge] add support for the JSON type"

+++ src/Symfony/Component/Messenger/Bridge/Redis/Transport/Connection.php
@@ -460,7 +460,7 @@ class Connection
                     throw new TransportException(json_last_error_msg());
                 }
 
-                $score = (int) ($this->getCurrentTimeInMilliseconds() + $delayInMs);
+                $score = $this->getCurrentTimeInMilliseconds() + $delayInMs;
                 $added = $this->connection->zadd($this->queue, ['NX'], $score, $message);
             } else {
                 $message = json_encode([
